test: finish TestFindIssues; try to implement AdjustMilestoneClient interface
